+++ .style.yapf
@@ -4,3 +4,4 @@ BLANK_LINES_AROUND_TOP_LEVEL_DEFINITION = 1
 BLANK_LINE_BEFORE_NESTED_CLASS_OR_DEF = false
 COLUMN_LIMIT = 99
 SPACES_BEFORE_COMMENT = 1
+INDENT_DICTIONARY_VALUE = true
