+++ grl/run.py
@@ -19,7 +19,7 @@ from grl.td_lambda import td_lambda
 from grl.policy_eval import PolicyEval
 from grl.memory import memory_cross_product, generate_1bit_mem_fns, generate_mem_fn
 from grl.pe_grad import pe_grad
-from grl.utils import pformat_vals, RTOL, results_path, numpyify_and_save
+from grl.utils import pformat_vals, results_path, numpyify_and_save
 from grl.mi import run_memory_iteration
 from grl.vi import value_iteration
