+++ scripts/hyperparams/pomdps_mi_dm.py
@@ -1,9 +1,8 @@
 hparams = {
     'file_name':
-    'runs_pomdps_mi_dm.txt',
+        'runs_pomdps_mi_dm.txt',
     'args': [{
-        'algo':
-        'mi',
+        'algo': 'mi',
         'spec': [
             'slippery_tmaze_5_two_thirds_up',
             'example_7',
@@ -15,18 +14,12 @@ hparams = {
             # 'bridge-repair', 'paint.95'
             # 'hallway'
         ],
-        'policy_optim_alg':
-        'dm',
-        'method':
-        'a',
-        'lr':
-        1,
-        'use_memory':
-        0,
-        'use_grad':
-        'm',
-        'mi_iterations':
-        2,
-        'seed': [2020 + i for i in range(10)]
+        'policy_optim_alg': 'dm',
+        'method': 'a',
+        'lr': 1,
+        'use_memory': 0,
+        'use_grad': 'm',
+        'mi_iterations': 2,
+        'seed': [2020 + i for i in range(10)],
     }]
 }
