+++ scripts/hyperparams/pomdps_mi_pi_q_abs.py
@@ -1,9 +1,8 @@
 hparams = {
     'file_name':
-    'runs_pomdps_mi_pi_q_abs.txt',
+        'runs_pomdps_mi_pi_q_abs.txt',
     'args': [{
-        'algo':
-        'mi',
+        'algo': 'mi',
         'spec': [
             'tmaze_5_two_thirds_up', 'example_7', 'tiger-alt', '4x3.95', 'cheese.95', 'network',
             'shuttle.95', 'paint.95'
@@ -11,25 +10,16 @@ hparams = {
             # 'bridge-repair',
             # 'hallway'
         ],
-        'policy_optim_alg':
-        'pi',
-        'value_type':
-        'q',
-        'error_type':
-        'abs',
-        'method':
-        'a',
-        'lr':
-        1,
-        'use_memory':
-        0,
-        'use_grad':
-        'm',
+        'policy_optim_alg': 'pi',
+        'value_type': 'q',
+        'error_type': 'abs',
+        'method': 'a',
+        'lr': 1,
+        'use_memory': 0,
+        'use_grad': 'm',
         # 'n_mem_states': [2, 3],
-        'n_mem_states':
-        2,
-        'mi_iterations':
-        2,
-        'seed': [2020 + i for i in range(10)]
+        'n_mem_states': 2,
+        'mi_iterations': 2,
+        'seed': [2020 + i for i in range(10)],
     }]
 }
