+++ scripts/hyperparams/tmaze_dm.py
@@ -1,6 +1,6 @@
 hparams = {
     'file_name':
-    'runs_tmaze_dm.txt',
+        'runs_tmaze_dm.txt',
     'args': [{
         'algo': 'mi',
         'spec': 'tmaze_5_two_thirds_up',
@@ -9,6 +9,6 @@ hparams = {
         'use_grad': 'm',
         'policy_optim_alg': 'dm',
         'mi_iterations': 0,
-        'seed': [2020 + i for i in range(30)]
+        'seed': [2020 + i for i in range(30)],
     }]
 }
