+++ scripts/hyperparams/tmaze_two_thirds_up_mem_grad.py
@@ -1,6 +1,6 @@
 hparams = {
     'file_name':
-    'runs_tmaze_two_thirds_up_mem_grad.txt',
+        'runs_tmaze_two_thirds_up_mem_grad.txt',
     'args': [{
         'algo': 'pe',
         'spec': 'tmaze_5_two_thirds_up',
@@ -8,6 +8,6 @@ hparams = {
         'use_memory': 0,
         'use_grad': 'm',
         'lr': 1,
-        'seed': [2020 + i for i in range(30)]
+        'seed': [2020 + i for i in range(30)],
     }]
 }
