+++ scripts/launching/run_cpu_locally.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/env bash
 export CUDA_VISIBLE_DEVICES=0
 
-cd ../
+cd ../../
 source venv/bin/activate
 parallel --eta -u < 'scripts/runs/runs_pomdps_mi_pi_q_abs.txt'
