+++ scripts/parse_alpha.py
@@ -1,11 +1,13 @@
+"""
+In this file we parse .alpha files generated by the program pomdp-solver from pomdp.org
+and create .npy files as described in grl/environment/pomdp_files/README.md
+"""
+
 import argparse
 import numpy as np
 from pathlib import Path
 
 from grl.environment.pomdp_file import POMDPFile
-"""
-In this file we parse .alpha files generated by the program pomdp-solver from pomdp.org
-"""
 if __name__ == '__main__':
     parser = argparse.ArgumentParser()
     parser.add_argument('filename', type=str, help='name of .alpha file to parse')
