+++ scripts/plotting/slippery_tmaze_mi_pi.py
